Fix the 64-bit build.
Canonical link: https://commits.webkit.org/27433@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@34990 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/kjs/JSLock.h
@@ -84,7 +84,7 @@ namespace KJS {
         static void lock(ExecState*);
         static void unlock(ExecState*);
 
-        static int lockCount();
+        static ssize_t lockCount();
         static bool currentThreadIsHoldingLock();
 
         static void registerThread();
@@ -98,7 +98,7 @@ namespace KJS {
             ~DropAllLocks();
             
         private:
-            int m_lockCount;
+            ssize_t m_lockCount;
             bool m_lockingForReal;
         };
     };
